+++ Sources/drstring-cli/Constants.swift
@@ -1,5 +1,5 @@
 enum Constants {
-    static let ignoreThrows = "ignoreThrows"
+    static let ignoreThrows = "ignore-throws"
     static let include = "include"
     static let exclude = "exclude"
 }