+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## master
+
+### New
+
+- Add subcommand 'check', which is the original command from 0.0.1
+- Add help messages for main command and `check`
+- Support option `--ignore-throws`, which makes DrString ignore throws in docstring validation.
+
 ## 0.0.1
 
 For SwiftPM projects, find existing docstring for function, report issues in these docstrings.